vm1(main_wms)/requirements.txt
Automat==0.8.0
cloud-init==24.4.1
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
dbus-python==1.2.16
distro==1.4.0
distro-info==0.23+ubuntu1.1
entrypoints==0.3
eventlet==0.39.1
Flask-Cors==5.0.0
Flask-SocketIO==5.5.1
GitPython==3.1.44
hyperlink==19.0.0
imutils==0.5.4
incremental==16.10.1
ipython==8.12.3
jaraco.collections==5.1.0
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
language-selector==0.1
launchpadlib==1.10.13
mysql-connector-python==9.0.0
netifaces==0.10.4
openai==0.27.0
pipdeptree==2.24.0
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyrsistent==0.15.5
pyserial==3.4
pytesseract==0.3.13
python-apt==2.0.1+ubuntu0.20.4.1
python-barcode==0.15.1
python-debian==0.1.36+ubuntu1.1
python-dotenv==1.0.1
pyzbar==0.1.9
pyzxing==1.0.2
-e git+https://github.com/xinntao/Real-ESRGAN.git@a4abfb2979a7bbff3f69f58f58ae324608821e27#egg=realesrgan
redis==5.2.1
reportlab==4.3.1
requests-unixsocket==0.2.0
sentry-sdk==2.25.1
service-identity==18.1.0
simplejson==3.16.0
sos==4.7.2
ssh-import-id==5.10
systemd-python==234
tensorboard==2.14.0
thop==0.1.1.post2209072238
torchaudio==2.1.2
Twisted==18.9.0
ubuntu-pro-client==8001
ufw==0.36
ultralytics==8.3.101
unattended-upgrades==0.1
zope.interface==4.7.1
